screenshotter/requirements.txt
bitstring==3.1.7
numpy==1.19.4
Pillow==8.0.1
